+++ src/main/java/com/prondzyn/fifadrawer/entities/NationalTeam.java
@@ -1,8 +1,10 @@
 package com.prondzyn.fifadrawer.entities;
 
+import java.math.BigDecimal;
+
 public class NationalTeam extends Team {
 
-  public NationalTeam(String name, Double rank) {
+  public NationalTeam(String name, BigDecimal rank) {
     this.name = name;
     this.rank = rank;
   }
